Patch: Parse append arguments properly in mboot module
Hey guys,

I noticed the appended ipappend options end up in the beginning of the
cmdline argv[] when mboot is entered. mboot then tries to parse the first
argument as the name of the kernel module and promptly fails cuz it can't
find "ip=.." or "bootif=...".

I'm attaching a patch which basically looks for any arguments in the argv[]
array and appends them to the kernel cmdline. The arguments have to be
back-to-back to be considered "appened options", so this should be safe
enough to not take out-of-place arguments and tack onto the kernel cmdline.

Last time i sent in some of the elf section loader patches, i think Peter
had trouble importing time. Hopefully it works now.

thanks,
Ram Yalamanchili