syslinux $PnP search fix
I finally fixed the "chain.c32 hangs when run on modern qemu" problem.
Apparently it was hanging hitting the "Invalid $PnP header (too short)" branch
of the code which loops without ever incrementing bx, so it never terminates.

The patch below (taken against your current git) seems to fix the problem for
me, but having not explored the code much, it probably isn't the most elegant
of fixes.

				-- Gerald