Tighten close_file, fix SYSLINUX new file API
Define close_file slightly tighter:

       - It is permitted to call close_file with a zero filehandle
       - close_file zeros SI

Fix the handling of the new API in SYSLINUX.  It appears file_left was
never actually initialized; it simply didn't matter since the old code
would either trip on reaching the end of the FAT chain or count the
data it needed.  Now operating according to spec.