The first functional prototype of module linking.
Although the code is currently a little messy, one
could see that when she loads hello_def and hello_ref
modules (in that order), the functions that
handle the shared symbol perform correctly when
tested by the main program. There is one initialization
function that stores 100 in the symbol, and a second one
that reads the symbol, returns its value and increments it.