isolinux: return failure when opening a zero-length file
In the Syslinux view of the world, a zero-length file doesn't exist
and therefore should return error on open.  A lot of the code relies
on this, but ISOLINUX allowed a zero-length file to be opened, which
could cause all kinds of trouble.

It would be nicer to not have that restriction, but removing it will
probably not happen until the configuration code is rewritten in C.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>