MEMDISK: in safeint mode, don't bother testing for unchanged INT 15h
In safeint mode, don't bother testing for unchanged INT 15h.  It
points back at ourselves, anyway, and we end up doing, in the end, the
same thing as the safeint stub would do.

Furthermore, the test was broken for years, and so this is actually
closer to what we originally tested.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>