Fix minor errors in keytab-lilo
H. Peter Anvin wrote:
> Thanks for the warning.  I have updated keytab-lilo in the git tree with
> keytab-lilo.pl from LILO 22.8.

Thank you H. Peter. I just checked it and the new version doesn't work
"out of the box". It says...
syntax error at keytab-lilo line 4, near "eval "
Execution of keytab-lilo aborted due to compilation errors.

Furthermore, after correcting lines 3 and 4 (which are responsible of
the above error), keyboard maps can't be found (by loadkeys utility)
because of an added bad extension (.map instead of .kmap).

Attached is a patch I wrote in order to make it fully work (apply it
against the version of keytab-lilo which is now in the git tree).