COM32 module: Read-Only shell
Well, here's the read-only shell (rosh) that I've been working on.
It's functional but still quite rough.  My primary intention of
posting it at this time is such that people have an easy way to
demonstrate to themselves that the library calls I made work.

This should apply as a patch to the head of the "dir" branch and the
patch for c_cflag/c_lflag.  The patch is only needed for Linux(without
it, segfaults will occur).  I've designed it to compile as a COM32
module and Linux binary when using GNU C (as Syslinux currently
requires GNU C, if I read the Makefiles correctly).