isohdpfx: work around gas/ld miscompilation
Some, unknown, version of gas or ld miscompiles a relative jump to an
absolute symbol, resulting in a broken isohybrid.  Implement a
workaround for it by using a far jmp, which is absolute.  Fortunately,
the isohybrid prefix is one of the very few boot sectors we have which
isn't short on space.