memdisk varieties: Allow for multiple memdisk installable hooks
This patch "sets the stage" for the memdisk kernel to include different
"installables": versions of the memdisk hooks.

This patch splits the case of the EDD-compiled hook from the non-EDD-compiled
hook.  Both are available inside the kernel, and the kernel chooses which one
to install based on the command-line.

It's possible that other, slightly modified versions of the memdisk hook could
be useful to have...