SYSLINUX: fix handling of .bss files for FAT12/FAT16
The size of the superblock is different for FAT12/FAT16 and FAT32.  We
should only copy the relevant portions of the superblock over.
Technically speaking, it is possible for FAT12/FAT16 to have an older,
smaller superblock as well (without the 0x29 boot signature); we don't
handle that at the moment.