memdisk: don't key FAT vs MBR on the type of disk
We can't rely on the disk type to say if this is FAT vs MBR, since we
haven't determined that yet.  Thus, move the FAT test up before the
MBR test; both are heuristic, but the FAT test has much less
probability of false positives, and can be further tightened if
needed.