hdt: Rewrite DMI module using a new syntax for the CLI
Impact: Start refactoring the CLI.

In order to have an extensible CLI, the commands handling needs to be
refactored. This patch starts the migration by refactoring the DMI
module.

Note that the core of the new commands parser has been implemented.
Adapting other modules to use it simply requires to declare
new structures that maps the command string to the callback.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>