hdt: Refactor help option; Cleanups
Impact: Help is usable in all modes, cleanups.

Help is now a generique helper that is shared by all modes.
It prints:
	* default commands for the current mode
	* the list of modules associated to show for the current mode
	* default commands for hdt

For the last item, it doesn't print overwritten commands, e.g. if the
dmi module implements a generic 'list' command and there is a generic,
shared, 'list' command in the hdt mode, the former overwrites the
latter.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>