hdt: Cleanups in the DMI module
Impact: Remove compilation warnings, Cleanups, DMI CLI output changed.

Add __unused macro to remove compilation warnings. argc and argv are
needed for consistency sake in the cli, but are not always used.

Slightly change the output of `list' to be more precise. It was
confusing for the first time user to list the modules with 'help' but
then have an error message when a module wasn't available.

Remove unused help function (there is a dynamic global one).

Misc.: move dprintf to the header file.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>