hdt: Convert syslinux module to use the new framework
Impact: Add the concept of default_callback per set of modules

Syslinux only implements a default show command. Previously, this was
triggered by the `show list' command line. It is not really intuitive
and may be confusing (it has another meaning in the DMI module).

There is now the concept of default callback, one that would be
triggered when only `show', `set', ... is entered in the command line.
A common usecase may be that `show' triggers a summary display.

Misc.: small fix in the DMI module.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>