shuffler: first cut of a simpler shuffle routine
Impact: major restructuring

New shuffler routine which runs entirely in protected mode, unlike the
old one which would enter PM for the bcopy phase only.  This is not
only faster, but it greatly reduces the size of the shuffler "critical
area".  The interfaces to this new shuffler are not yet implemented,
so this is a non-functional checkpoint.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>