com32: make memory beyond the core HighMem available to malloc
Impact: should deal with memory holes more gracefully

Right now, if we find a memory hole, we simply don't use the memory
beyond that point.  This makes it possible for com32 modules to
use that memory.  After this, we should be able to add relocation
support to allow loading when the memory at 1 MB is unavailable.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>