core: don't misalign xfer_buf_seg
xfer_buf_seg is used for bulk transfers, and we really don't want
to break it up in order to avoid 64K DMA boundaries.

Note: the same is true for the com32 bounce buffer, which is
real_mode_seg.  However, there are hard assumptions in the code that
real_mode_seg is the last segment; those need to be cleaned up, or we
need to verify that xfer_buf_seg can be used as the com32 bounce
buffer.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>