bcopyxx: when going to 16-bit PM, might as well do it right
When entering 16-bit PM after shuffle and boot, we might as well
do so sanely.  Specifically, set up the data segments so that they
match the code segment, generating a 16-bit "tiny" model environment.

This makes it a lot saner to bootstrap a proper PM environment from
there if that is what the user intends.  For the presumably more
common case of RM entry, it won't do any harm, and it's only a handful
of additional instructions.

Signed-off-by: H. Peter Anvin <hpa@linux.intel.com>