mboot: reimplement the Solaris DHCP hack, add compliant a.out mode
Reimplement the Solaris DHCP information passing hack.

Add a spec-compliant mode for the "a.out kludge".  The spec is pretty
clear that the bit should override the ELF header (after all,
otherwise there wouldn't be any need for the bit), but Grub
disagrees.  We default to Grub-compliant mode, as Solaris seems to set
the bit even though it's an ELF kernel, but add the option to enable
spec-compliant mode, as apparently some versions of FreeBSD need it.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>