bcopyxx: clear all flags before jumping to the successor code
Clear the flags register while we still have a stack to do so.
This makes the state in the target OS more predictable.  It's worth
noting that none of the instructions we use in either the PM or RM
trampolines change the flags, either (except the IF setting in the RM
trampoline, but that is intentional.)

Signed-off-by: H. Peter Anvin <hpa@zytor.com>