Unify dependency generation
Make the dependency generation more common; have a general pattern in
MCONFIG, and use it in rules (not in CFLAGS).

For NASM source, in order to stay compatible with old versions of
NASM, run NASM twice; newer versions of NASM is capable of generating
dependencies simultaneously like gcc can, but that would break
compatibility with older distros.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>