ADV: EDD "write with verify" is AL=01h or AL=02h, not AL=80h
The "write with verify" functionality in EDD is AL=01h or AL=02h
depending on EDD version, not AL=80h.  Since it is apparently
version-dependent, don't even try to do it; furthermore, some BIOSes
are known to reject a write with this feature requested no matter
what.  At some point we may want to explicitly read back to verify.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>