hdt: Silence harmless errors when detecting disks
Detecting disk geometry can fail either because the extension query failed
(invalid disk number?) or because the legacy geometry call failed
(int 13 not supported?). The latter happens for instance when querying
geometry for CDROMs.

In either case, the error message is not meaningful. Don't print it (and
don't confuse the user).

(Impact: fixes #5)

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>