Don't run to comboot API with interrupts off
We have historically run the comboot API with interrupts off due to
concerns about stack overflow.  However, this is a really bad idea: we
can easily spend a fair bit of time servicing one of these routines,
especially when doing things like waiting for I/O.

In particular, do_idle should *never* be run with interrupts disabled.

Switch to running with interrupts enabled everywhere unless we have
specific reasons not to do so.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>