pxelinux: revert to a separate stack implementation
Return to using a private stack, except immediately around a pxenv
call.  The systems on which this broke has been lost in time, and with
more code in 32-bit space there is a risk of deeper stacks (due to
more mode-switch calls.)

It is still a single configuration variable to flip back to the
unified 16-bit stack, however.  If most of the code is in 32-bit mode
and there is small likelihood of several transitions, it might still
be the best option in the end.

Note: the PXE stack is 1.5K in size (and we must account for the
possibility of receiving an interrupt.)

Signed-off-by: H. Peter Anvin <hpa@zytor.com>