isolinux: handle a core file greater than 64K
Handle a core file greater than 64K, and handle the case of a BIOS
which can't handle 64K wraparounds.  The latter is messy with CD-ROMs,
since the standard load address (0x7c00) is not 2K aligned.  We will
have to do something similar for disk-based derivatives when dealing
with 4K sector drives.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>