gpllib: Add support for 3.3.11 On Board Devices Information (Type 10, obsolete) (dmi)
From SMBIOS spec (v2.6.1):

The information in this structure defines the attributes of devices that are
onboard (soldered onto) a system element, usually the baseboard. In general, an
entry in this table implies that the BIOS has some level of control over the
enabling of the associated device for use by the system.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>