Convert the cache code to C and implement the core printf function
for the cahce part, I do get the error message says that undefined reference
to `getlinsec'. I'm abort to implement a C version one.

for the printf function, it works somehow, but doesn't work well. With the test,
it seems it can handle the format output correctly. And I haven't debugged it,
so I have no idea for now.