Convert the config load part to C
For now, I added a temporary file load_config.c that would be merged into
extlinux.c file later, which would make my later work be much easier.

And in order to get rid of the conflict noisy message from the open function
in ui.inc and the open function defined in the unistd.h, I changed the
open in core to core_open. It may be ugly, but it works :)