simple menu: make ONTIMEOUT work with MENU HIDDEN
The command invoked when no keys were pressed for the specified time
(ONTIMEOUT) may be different from the menu item initially selected
when the menu is displayed (DEFAULT or MENU DEFAULT).  Unfortunately,
this did not work together with MENU HIDDEN (which is exactly the case
when having a separate ONTIMEOUT command makes the most sense).

Signed-off-by: Sergey Vlasov <vsu@altlinux.ru>
Signed-off-by: H. Peter Anvin <hpa@zytor.com>