Core: Convert the searchdir function to C
for now, almost all the stuff related to EXT have been converted to C,
and fow my (limit) test, it works well.

The getfssec function haven't converted to C, because it also be called
from asm file, and I find it's a bit hard to convert it to C. But however,
it's my next plan.