Core:EXTLINUX: fix bug with lots of dots and searchdir
The bug hpa said in the email that prints lots of dots when loading a kernel fixed.
And with my test, I found that the searchdir couldn't return correctly while not find
what we wanna find.(In asm, the asm code use ZF flag to test if we successed. And I
thought regs->esi.w[0] = file(be zero on fail) will set ZF flag). Well, It's a good
thing that I found the reg structure has the eflags filed. then it works; nice thing!