Core:EXTLINUX: applies the path from hpa to EXTLINUX
I just added it to EXTLINUX first, and I will consider if I should added it
SYLSINUX and ISOLINUX and so on tomorrow (I'm not sure for now).

And I added a new structure disk to store the disk specific information, so
some changes happened on the vfs-like interface. Hope you will like it, hap;)

the new code broken on the non-standard disk geometry, like the exttest package
given by hpa.