cli: edit_cmdline() that actually works
Fix a number of issues with edit_cmdline when used in a
scrolling-screen context.  Here we are forced to keep track of
coordinates within our chunk of the screen; unfortunately this is not
generic w.r.t. window width on the serial console, but it should at
least behave properly within the given width.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>