cmenu: use VT-100 alternate character set for drawing boxes
Use the Special Characters and Line Drawing Character Set as
described in http://www.vt100.net/docs/vt102-ug/table5-13.html
to draw boxes around the menus.

Note that to work in xterm, G1 needs to be initialized (to point
to the alternate character set). This is done in the cls function
("\033)0").

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>