cmenu: reset attributes instead of using ANSI sequence 22
hpa advised not to use ANSI sequence 22, as it may not work everywhere.
Instead, let's reset attributes and set last_attr to unknown. The next
background/foreground tests are then guaranteed to fire. We take a small
hit in terms of performance though.

Misc.: change initialization of last_attr to 0x300 (unknown) to trigger
properly the tests the first time.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>