gpllib: share PAGE_SIZE definition with cpuid.h
swsusp detection needs PAGE_SIZE. cpuid.h is actually already defining it.
Since both swsusp and cpuid are taken from the Linux kernel, it is a good idea to
share it (they really need to be equal).

We are going to rely on cpuid.h. Creating a page.h for this single definition seems
overkilling and the MMU is part of the CPU on x86 anyways.

Impact: fix compiler warning (redefinition).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>