Use jiffies() to implement times(); create <core/*.h> namespace
Use the core function jiffies() to implement times(), instead of
poking at the BIOS variable.

This creates a new need for a include file namespace for items
exported from the core for the purpose of the library (not intended
for direct use by applications).  Use the <core/*.h> namespace for
that.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>