core: add a way to invoke a function on return to protmode
Add a hook so we can invoke a protmode function immediately before
enabling interrupts on the way back from protected mode.  This will
*not* be invoked if we are more than one level deep (i.e.
[RM(0) ->] PM(0) -> RM(1) -> PM(1) -> RM(2) will not invoke the hook
when returning to PM(1).)  This is intended to provide a hook for
taking an action after a real-mode interrupt did something, i.e. the
PXE receive ISR or the timer tick.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>