libansi: reset SGR parameters cache in cls()
There was a bug in HDT when switching from the menu to the cli and
back to the menu: the background wasn't redrawn.

This was due to the fact that the cli bypasses the libansi library
and calls printf() directly. Unfortunately, the SGR attributes were
cached between menu initializations and the background wasn't properly
redrawn.

To fix it, invalidate the cache in the cls() routine. cls() does much
more than just erasing the screen already (e.g. initialize G1) - we
may want to rename the function at some point.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>