hdt: Adding automatic mode
Impact: HDT can be controlled via the cmdline

When the auto= parameter is set, HDT execute a set of commands and
return to the cli after.

i.e auto='show summary; show pci' will show the summary then the pci and
return to the cli.

Note that the auto= feature, requires a the "'" char to define the
complete set of commands and the ";" char to separate commands.

Some more examples:
auto='show syslinux; exit' -> this will show the syslinux release and
then exit to return to the bootloader

auto='show summary; reboot' -> this will show the summary and
then reboot the computer