FS: add a new fs flag and do a better searchdir-check
If we want use the malloc/free functions, set the FS_USEMEM flag
in the (fs)_ops.flags, and in the fs_init phase, it will init the
memory management system.

And it would be little better to do a check base on if we have searchdir
method or not instead of the fs name in the searchdir function.

Signed-off-by: Liu Aleaxander <Aleaxander@gmail.com>