memory: Adding memory size detection
Impact: allow com32 module to detect the amount of memory installed

This commit adds several function:

sanitize_e820_map() to generate a new e820 map that doesn't have
overlaping

memsize_e820() to detect the available memory (in KB)

detect_memsize() to detect the amount of memory by using e820/e801/e88.
It first try to detect the memory via e820. If it fails, it uses e801 as
fallback. If this e801 fails, e88 is used.