memdisk: "safe hook" and mBFT
Two additions to MEMDISK to support OS drivers.

The "safe hook" structure ("Safe Master Boot Record INT 13h Hook Routines")
is a means for an OS driver to follow a chain of INT 13h hooks, examining
the hooks' vendors and assuming responsibility for hook functionality along
the way.  For MEMDISK, we guarantee an additional field which holds the
physical address for the mBFT.

The mBFT is an ACPI table which an OS driver can scan for.  The mBFT
contains the official MEMDISK Info structure (MDI) which itself includes
parameters the OS will want to know about.  The mBFT points back at the
"safe hook" structure's physical address so that an OS supporting both
"safe hook" chain-walking as well as mBFT-scanning can know that both
refer to the same MEMDISK instance.

Signed-off-by: Shao Miller <shao.miller@yrdsb.edu.on.ca>
Signed-off-by: H. Peter Anvin <hpa@zytor.com>