LDLINUX:vfat: Improve the fat fs relative path searching
Added a new flag, FS_THISIND, used in the fat and iso fs relative path
searching. It means tries to get the last directory of the config file
path in the 'this_inode' variable to do a relative searching while loading
the kernel images in after.

And also fixed a bug in fs.c which wouldn't close the file structure when
doing path searching failed! And as a result, we could just run a few com32
programs.

Signed-off-by: Liu Aleaxander <Aleaxander@gmail.com>