LDLINUX:vfat: Oh, NO, I did fix the above bug in a wrong way!
I shouldn't count on the 'dname' parameter, we should ZERO the
'long_name' buffer at every time we met a new long name entry,
and that's the right way to fix the bug.

And we shouldn't count the 'id == 0' to check if we have a long
name matched first or not. Say we have one entry of a long name
entry, it will always be 'id == 0' even it doesn't matches. So,
add a new flag variable to do this.

Signed-off-by: Liu Aleaxander <Aleaxander@gmail.com>