pxe: update to match the new fs_init interface
fs_init now takes a pointer to an array of filesystems, so provide it
as such.  Furthermore, fs_init needs to return a nonnegative value to
be considered successful.

Finally, handle the case of 64-bit partition offsets being passed in
to fs_init (in ebx:ecx in the disk-based derivatives.)

Signed-off-by: H. Peter Anvin <hpa@zytor.com>