core: initial work on path-based cwd selection
Work on picking the initial cwd by storing a path instead of by
storing an inode number.  This should be both more general (in the
sense of supporting filesystems in a generic way) as well as
conceptually cleaner.  The code doesn't work yet, but this at least
provides support for the extlinux installer to store its subpath into
the installed image.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>